fix: change data type from numeric to double

+++ pg/scripts/init-schema.sql
@@ -8,8 +8,8 @@ create table enzymes (
     reaction_id char(6),
     ec_number varchar(255) not null,
     compound char(6),
-    kcat_values numeric,
-    km_values numeric
+    kcat_values float(53),
+    km_values float(53)
 );
 
 create table compounds (
